Fix the integer division.
The / operator in Python 3 represent float division, so we have to convert the result to int().

+++ software/models/stochasticModel.py
@@ -25,7 +25,7 @@ def stochasticModelAnal(x, H, N, stocf):
 	if not(UF.isPower2(N)):                                # raise error if N not a power of two
 		raise ValueError("FFT size (N) is not a power of 2")
 		
-	hN = N/2                                                # half of fft size
+	hN = int(N/2)                                           # half of fft size
 	w = hanning(N)                                          # analysis window
 	x = np.append(np.zeros(hN),x)                           # add zeros at beginning to center first window at sample 0
 	x = np.append(x,np.zeros(hN))                           # add zeros at the end to analyze last sample
@@ -50,7 +50,7 @@ def stochasticModelSynth(stocEnv, H, N):
 	returns y: output sound
 	"""
  
-	hN = N/2                                                 # half of FFT size for synthesis
+	hN = int(N/2)                                            # half of FFT size for synthesis
 	L = stocEnv[:,0].size                                    # number of frames
 	ysize = H*(L+3)                                          # output sound size
 	y = np.zeros(ysize)                                      # initialize output array
@@ -88,7 +88,7 @@ def stochasticModel(x, H, N, stocf):
 	if not(UF.isPower2(N)):                                  # raise error if N not a power of twou
 		raise ValueError("FFT size (N) is not a power of 2")
 		
-	hN = N/2                                                 # half of FFT size for synthesis
+	hN = int(N/2)                                            # half of FFT size for synthesis
 	w = hanning(N)                                           # analysis/synthesis window
 	x = np.append(np.zeros(hN),x)                            # add zeros at beginning to center first window at sample 0
 	x = np.append(x,np.zeros(hN))                            # add zeros at the end to analyze last sample
