+++ software/models/utilFunctions.py
@@ -144,7 +144,7 @@ def genSpecSines_p(ipfreq, ipmag, ipphase, N, fs):
 	"""
 
 	Y = np.zeros(N, dtype = complex)                 # initialize output complex spectrum  
-	hN = N/2                                         # size of positive freq. spectrum
+	hN = int(N/2)                                    # size of positive freq. spectrum
 	for i in range(0, ipfreq.size):                  # generate all sine spectral lobes
 		loc = N * ipfreq[i] / fs                       # it should be in range ]0,hN-1[
 		if loc==0 or loc>hN-1: continue
@@ -320,7 +320,7 @@ def sineSubtraction(x, N, H, sfreq, smag, sphase, fs):
 	returns xr: residual sound 
 	"""
 
-	hN = N/2                                           # half of fft size
+	hN = int(N/2)                                      # half of fft size
 	x = np.append(np.zeros(hN),x)                      # add zeros at beginning to center first window at sample 0
 	x = np.append(x,np.zeros(hN))                      # add zeros at the end to analyze last sample
 	bh = blackmanharris(N)                             # blackman harris window
@@ -351,7 +351,7 @@ def stochasticResidualAnal(x, N, H, sfreq, smag, sphase, fs, stocf):
 	returns stocEnv: stochastic approximation of residual 
 	"""
 
-	hN = N/2                                              # half of fft size
+	hN = int(N/2)                                         # half of fft size
 	x = np.append(np.zeros(hN),x)                         # add zeros at beginning to center first window at sample 0
 	x = np.append(x,np.zeros(hN))                         # add zeros at the end to analyze last sample
 	bh = blackmanharris(N)                                # synthesis window
